fetch/requirements.txt
Pillow==10.0.0
requests==2.28.2
